requirements.txt
pdfplumber>=0.11.0
google-generativeai>=0.8.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pyinstaller>=6.0.0
